feat(js-debugger): Refactor command following SEAMS and DRY principles
Major architectural improvements:
- Split 800+ line command.ts into thin orchestrator (~300 lines)
- Created SEAMS interfaces for extensibility (IToolHandler, IResponseFormatter, ISessionValidator)
- Applied DRY principles: extracted session validation, response formatting, mock logic
- New modular structure: handlers/, formatters/, sessions/, adapters/
- Separated mock and real implementations
- Added factory pattern for handler registration
- Created 8 individual tool handlers replacing monolithic switch statements
- Consolidated JSON response formatting to eliminate duplication
- Improved maintainability and testability
