+++ packages/commands/js-debugger/src/formatters/index.ts
@@ -1,0 +1,2 @@
+export { DebugResponseFormatter } from './debug-response-formatter.js';
+export type { IResponseFormatter } from '../types.js';
