+++ packages/commands/js-debugger/src/handlers/index.ts
@@ -1,0 +1,20 @@
+export { DebugHandler } from './debug-handler.js';
+export { ContinueHandler } from './continue-handler.js';
+export { ListSessionsHandler } from './list-sessions-handler.js';
+export { StopHandler } from './stop-handler.js';
+export { GetStacktraceHandler } from './get-stacktrace-handler.js';
+export { GetVariablesHandler } from './get-variables-handler.js';
+export { SearchConsoleOutputHandler } from './search-console-output-handler.js';
+export { CleanupSessionsHandler } from './cleanup-sessions-handler.js';
+
+export type { IToolHandler, ToolHandlerContext } from '../types.js';
+
+// Re-export all handler argument types for convenience
+export type { DebugHandlerArgs } from './debug-handler.js';
+export type { ContinueHandlerArgs } from './continue-handler.js';
+export type { ListSessionsHandlerArgs } from './list-sessions-handler.js';
+export type { StopHandlerArgs } from './stop-handler.js';
+export type { GetStacktraceHandlerArgs } from './get-stacktrace-handler.js';
+export type { GetVariablesHandlerArgs } from './get-variables-handler.js';
+export type { SearchConsoleOutputHandlerArgs } from './search-console-output-handler.js';
+export type { CleanupSessionsHandlerArgs } from './cleanup-sessions-handler.js';
