+++ packages/commands/js-debugger/src/handlers/list-sessions-handler.ts
@@ -1,0 +1,36 @@
+import type {
+  IToolHandler,
+  ToolHandlerContext,
+  CallToolResult,
+} from '../types.js';
+
+export interface ListSessionsHandlerArgs {
+  // No arguments needed for listing sessions
+}
+
+/**
+ * Handler for listing active debug sessions
+ * Implements the IToolHandler SEAM for modular tool handling
+ */
+export class ListSessionsHandler
+  implements IToolHandler<ListSessionsHandlerArgs>
+{
+  readonly name = 'list_sessions';
+
+  async handle(
+    _args: ListSessionsHandlerArgs,
+    context: ToolHandlerContext,
+  ): Promise<CallToolResult> {
+    try {
+      const realSessions = context.sessionManager.listSessions();
+      const mockSessions = context.mockSessionManager?.listMockSessions() || [];
+
+      return context.responseFormatter.sessionsList(realSessions, mockSessions);
+    } catch (error) {
+      return context.responseFormatter.error(
+        error instanceof Error ? error.message : 'Unknown error',
+        { operation: 'list_sessions' },
+      );
+    }
+  }
+}
