+++ packages/commands/js-debugger/src/sessions/index.ts
@@ -1,0 +1,2 @@
+export { SessionValidator } from './session-validator.js';
+export type { ISessionValidator } from '../types.js';
